initial changes
add the contact us file to the button so that when user clicks on the button it redirrects tothe contact us file.

Also solve issue which was overlaping of the contact us input tags when user scrolls up

Add a box for the information i.e (phone numbers,social media and etc).